Rework Semaphore to return a SemaphoreGuard
This is much better than requiring users to manually call `Semaphore::release()`
but I'm still not perfectly satisfied with the API because, despite the names,
both `Semaphore::release()` and `Semaphore::try_release()` are not exactly safe
- only that they _may_ catch some cases of incrementing a semaphore's
availability count past its maximum.

The problem occurs when an availability count is decremented because a semaphore
has been acquired, at which point it becomes possible to "safely" call
`Semaphore::try_release(n)` where `n = max_count - current_count`. This won't
panic (and `try_release()` won't return an error) but the next call to
`Semaphore::release()` when that temporarily obtained semaphore is returned
*would* panic, because now the count is invalid.