Drop Semaphore count to 16-bit variable
Since we now need to include both the "max" and the "current max" in our
semaphore in addition to the core "current count" field, lets try dropping the
backing size from a 32-bit integer to a 16-bit integer.

That's still a maximum of 65536 concurrency slots, which seems... sufficient?
We'll RFC when this is released alongside the other concerns encountered.