Revert `&mut` requirement for `CountdownEvent::reset()`
After some consideration, if it's safe to call `AtomicU8::store()` even though
other threads might be reading/writing to the same location then there's no
reason `CountdownEvent::reset()` should be any different.

Unlike our issue with the previous version of `Semaphore::release()`, this
operation can't invalidate any internal preconditions, contracts, or assumptions
held by the countdown event and so there's no reason to forbid it just because
someone might use it in a way that doesn't yield the results _they_ expect (so
long as it yields the results _we_ expect, and it does).