Add `CountdownEvent::increment()` and make `reset()` non-racy
To improve the utility of `CountdownEvent` and make it applicable for a wider
range of applications, a new member function `CountdownEvent::increment()` has
been added that is the opposite of `CountdownEvent::tick()` and increments the
number of outstanding work items/objects/whatever tracked by the internal
counter.

To make this non-racy such that simultaneous calls to `tick()` and `increment()`
while the count was at 1 or 0 do not result in a missed wake nor do they result
in the event being incorrectly set/available, a second synchronization object
(a 1-byte `AutoResetEvent`) has been added.

Acquiring this event is ~free so long as there are no simultaneous calls between
(`increment()` or `reset()`) and `tick()` (or simultaneous calls to
`increment()` and `reset()`), making it basically free for the common use cases.

With this change, `CountdownEvent::reset()` is now also non-racy and can be used
at any time with correct semantics.